% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{browse_exercise}
\alias{browse_exercise}
\title{Navigate to an exercise description on exercism.io}
\usage{
browse_exercise(track_id = "r", slug)
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}

\item{slug}{a normalized, url-safe identifier for a problem
e.g. "hello-world"}
}
\description{
Navigate to an exercise description on exercism.io
}
\examples{
\dontrun{
browse_exercise("r", "leap")
}
}
