% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{browse_solution}
\alias{browse_solution}
\title{Navigate to an exercise solution on exercism.io}
\usage{
browse_solution(track_id = "r", slug)
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}

\item{slug}{a normalized, url-safe identifier for a problem
e.g. "hello-world"}
}
\description{
Navigate to an exercise solution on exercism.io
}
\examples{
browse_solution("r", "hamming")
}
