% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_next_problem}
\alias{check_next_problem}
\title{Check the next problem for a particular language track}
\usage{
check_next_problem(track_id = "r")
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track. e.g. r, python, javascript etc}
}
\value{
Prints out current/next problem, silently returns problem slug
}
\description{
Returns the next problem for a language track i.e. the next problem to be
submitted, or possibly the next problem to be fetched
}
