% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_next}
\alias{fetch_next}
\title{Fetch the next problem for a particular language track}
\usage{
fetch_next(track_id = "r", skip = FALSE, force = FALSE, open = TRUE,
  run_tests = FALSE)
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}

\item{skip}{logical, indicating whether to skip the current (unsubmitted)
problem and fetch the next problem}

\item{force}{logical, indicating whether existing problem files should be
overwritten}

\item{open}{logical, indicating whether to open the problem files immediately
after fetching}

\item{run_tests}{logical, indicating whether to run \code{testthat::auto_test()}
on the fetched problem's directory}
}
\value{
Prints confirmation message upon success
}
\description{
Checks for the next problem via the Exercism API, and writes the files into
the folder in the Exercism directory
}
\examples{
\dontrun{
fetch_next("r")
}
}
