% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_problem}
\alias{fetch_problem}
\title{Fetch a problem for a particular language track}
\usage{
fetch_problem(trackID = "r", slug, force = FALSE)
}
\arguments{
\item{trackID}{a normalized, url-safe identifier for a language track. e.g. r, python, javascript etc}

\item{slug}{a normalized, url-safe identifier for a problem e.g. "hello-world"}

\item{force}{logical, indicating whether existing problem files should be overwritten}
}
\value{
Prints confirmation message upon success
}
\description{
Fetches the files for a problem via the Exercism API and writes them into
a new problem folder in the Exercism directory
}
