% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_exercise}
\alias{open_exercise}
\title{Open files for an exercism.io problem}
\usage{
open_exercise(slug, track_id = "r", change_wd = FALSE)
}
\arguments{
\item{slug}{a normalized, url-safe identifier for a problem
e.g. "hello-world"}

\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}

\item{change_wd}{logical, indicating whether the working directory should be
set to the directory of the exercise being opened}
}
\description{
Open files for an exercism.io problem
}
\examples{
dontrun{
open_exercise("leap")
}
}
