% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.status}
\alias{print.status}
\title{Print status object}
\usage{
\method{print}{status}(x, ...)
}
\arguments{
\item{x}{a status object (returned via \code{\link{track_status}}).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print status object
}
