% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key for exercism.io}
\usage{
set_api_key(key, force = FALSE)
}
\arguments{
\item{key}{API key for exercism.io}

\item{force}{logical, indicating whether to allow overwriting of an existing
EXERCISM_KEY environment variable}
}
\value{
Silently returns the result of the set operation, which will either
be 'Updated' an existing key or 'Created' a new key.
}
\description{
Set an environment variable for the provided exercism.io API key, and store
in .Renviron so that it can persist for future sessions.
}
