% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_exercism_path}
\alias{set_exercism_path}
\title{Set path for exercism}
\usage{
set_exercism_path(path = NULL, force = FALSE)
}
\arguments{
\item{path}{path to exercism folder}

\item{force}{logical, indicating whether to allow overwriting of an existing
EXERCISM_PATH environment variable}
}
\value{
Silently returns the result of the set operation, which will either
  be 'Updated' an existing key or 'Created' a new key.
}
\description{
Set an environment variable for the provided exercism path, and store
in .Renviron so that it can persist for future sessions.
}
