% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{status}
\alias{status}
\title{Get status from exercism.io}
\usage{
status(track_id = "r")
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track. e.g. r, python, javascript etc}
}
\value{
Current status from exercism.io
}
\description{
Fetches current status from exercism.io
}
