% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit}
\alias{submit}
\title{Submit a solution to exercism.io}
\usage{
submit(path, comment = NULL, browse = FALSE)
}
\arguments{
\item{path}{full path to the file containing the solution}

\item{comment}{comment to include with the submission}

\item{browse}{logical value indicating whether to navigate to the submission
on exercism.io on completion}
}
\value{
Response from exercism.io
}
\description{
Submits the specified solution to exercism.io
}
\examples{
submit("~/exercism/r/hello-world/hello-world.R")
}
