% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{track_status}
\alias{track_status}
\title{Get track status from exercism.io}
\usage{
track_status(track_id = "r")
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}
}
\value{
Current track status from exercism.io
}
\description{
Fetches current track status from exercism.io
}
