% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_api_response}
\alias{check_api_response}
\title{Check API response}
\usage{
check_api_response(response)
}
\arguments{
\item{response}{An http response from GET or POST}
}
\value{
Doesn't return a value, simply throws a warning or error if there
is something wrong with the response.
}
\description{
Checks the http response code and type.
}
