% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.status}
\alias{print.status}
\title{Print status object}
\usage{
\method{print}{status}(status, ...)
}
\arguments{
\item{status}{a status object (returned via \code{\link{status}}).}
}
\description{
Print status object
}
