% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run tests against your solution automatically}
\usage{
run_tests(slug, track_id = "r")
}
\arguments{
\item{slug}{a normalized, url-safe identifier for a problem
e.g. "hello-world"}

\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}
}
\value{
Prints test reports
}
\description{
Exercism-specific wrapper for \code{\link[testthat:auto_test]{testthat::auto_test()}} that starts running the
problem's test cases against your solution. Thus, you can improve it iteratively, in a "\href{http://exercism.io/how-it-works/newbie}{test-driven}" manner.
}
\examples{
\dontrun{
  run_tests("hello-world")
}
}
