% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{skip_problem}
\alias{skip_problem}
\title{Skip a problem for a language track}
\usage{
skip_problem(track_id = "r", slug)
}
\arguments{
\item{track_id}{a normalized, url-safe identifier for a language track.
e.g. r, python, javascript etc}

\item{slug}{a normalized, url-safe identifier for a problem
e.g. "hello-world"}
}
\value{
Prints confirmation message upon success
}
\description{
Marks a problem as 'skipped' via the Exercism API
}
